Module ffi_support::error

Structs

ErrorCode
A wrapper around error codes, which is represented identically to an i32 on the other side of the FFI. Essentially exists to check that we don't accidentally reuse success/panic codes for other things.

ExternError
Represents an error that occured within rust, storing both an error code, and additional data that may be used by the caller.